requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pandas==2.1.4
python-binance==1.0.19
pillow==10.1.0
schedule==1.2.0